+++ Elements/src/Geometry/Curve.cs
@@ -1,12 +1,13 @@
 using System.Collections.Generic;
 using Elements.Geometry.Interfaces;
+using Newtonsoft.Json;
 
 namespace Elements.Geometry
 {
     /// <summary>
     /// The abstract base class for all curves.
     /// </summary>
-    [Newtonsoft.Json.JsonConverter(typeof(Elements.Serialization.JSON.JsonInheritanceConverter), "discriminator")]
+    [JsonConverter(typeof(Elements.Serialization.JSON.JsonInheritanceConverter), "discriminator")]
     public abstract partial class Curve : ICurve, ITransformable<Curve>
     {
         /// <summary>
@@ -96,9 +97,9 @@ namespace Elements.Geometry
         /// <param name="c">The curve to convert.</param>
         public static implicit operator ModelCurve(Curve c) => new ModelCurve(c);
 
-        internal GraphicsBuffers ToGraphicsBuffers(bool lineLoop)
+        internal GraphicsBuffers ToGraphicsBuffers()
         {
-            return this.RenderVertices().ToGraphicsBuffers(lineLoop);
+            return this.RenderVertices().ToGraphicsBuffers();
         }
     }
 }